refactor: split userspace manager snapshot builders
